requirements.txt
faster-whisper>=1.0.0
torch==1.13.1
torchaudio==0.13.1
pyannote.audio==2.1.1
pydub>=0.25.1
ffmpeg-python>=0.2.0
numpy==1.26.4
pandas>=2.0.0
scikit-learn>=1.2.0
rapidfuzz>=2.14.0
unidecode>=1.3.6
nltk>=3.8.1
tqdm>=4.65.0
python-dateutil>=2.8.2
